+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 * April 4, 2022 | 11:54 PM
-Removed minimum line length so that code can be generated without even needing a prompt (Just from context)
+Removed minimum line length so that code can be generated without even needing a prompt (Just from context) \
 Added support for python comments
+* April 5, 2022 | 11:33 AM
+Created the discord bot to view the report status of CodeXR \
+Added statusData to the database and added it to the back-end